Simplfy processing configuration
We need to reduce the amount of processing configuration done on an
ad-hoc basis in a scenario. At-least until it can be removed
entirely.

This change simplifies the encryption tests to reduce the chances
of failure and make them as consistent as possible.

This change should bring with it minor speed improvements as well.